Add configure to extension
